% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{ARS}
\alias{ARS}
\title{Generate Adjusted R Squared}
\usage{
ARS(Target, Prediction)
}
\arguments{
\item{Target}{A vector of the dependent variable.}

\item{Prediction}{A vector of a covariate.}
}
\description{
This function computes the Adjusted R Squared equivalent to that obtained
from a linear regression.
}
\examples{
ARS(mtcars$mpg, mtcars$cyl)
}
