% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUROC.R
\name{AUROC}
\alias{AUROC}
\title{Area Under the ROC Curve}
\usage{
AUROC(Target, Prediction)
}
\arguments{
\item{Target}{A vector with values of 0/1 to be predicted.}

\item{Prediction}{A vector of prediction or covariate used to predict the target.}
}
\description{
This function computes the AUC, and returns the numeric value of it.
}
\examples{
AUROC(mtcars$am, mtcars$mpg)
}
