% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalTable.R
\name{CategoricalTable}
\alias{CategoricalTable}
\title{Cross Table of Categorical Covariates}
\usage{
CategoricalTable(Target, Covariate, Data, UseLogit = T)
}
\arguments{
\item{Target}{The name of binary target to be predicted.}

\item{Covariate}{The name of the covariate.}

\item{Data}{A data.frame or data.table containing both the target and covariate.}

\item{UseLogit}{If the value is TRUE, Log Odds will be generated. Otherwise,
a set of score derived from Log Odds, scaled from 0 to 100, will be generated.}
}
\description{
This function generates a data.table tabulating the different levels of a categorical
independent variable and the corresponding probabilities of the binary dependent
variable taking the value of 1.
}
\examples{
CategoricalTable(Target = "am", Covariate = "gear", Data = mtcars)
}
