% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cut.R
\name{Cut}
\alias{Cut}
\title{Cut}
\usage{
Cut(Data, NewColumn, x, breaks, include.lowest = T, ...)
}
\arguments{
\item{Data}{The data.frame or data.table for cut() to be used in.}

\item{NewColumn}{The name for the new column.}

\item{x}{Argument for cut(), a numeric vector which is to be converted to a
factor by cutting.}

\item{breaks}{Argument for cut(), either a numeric vector of two or more
unique cut points or a single number (greater than or equal to 2) giving the
number of intervals into which x is to be cut.}

\item{include.lowest}{Argument for cut(), logical, indicating if an ‘x[i]’
equal to the lowest (or highest, for right = FALSE) ‘breaks’ value should be
included.}

\item{...}{Other argument for cut().}
}
\description{
This function discretizes a column of numerical values and returns the
values in a new column.
}
\details{
It is similar to base::cut() but it also arranges the new column to be right
next to the old column being discretized.
}
\examples{
Discretize a continuous covariate
mtcars[, mpgCat := cut(mpg, breaks = c(10, 17, 21, 35), include.lowest = T)]

is comparable to

Cut(Data = mtcars, NewColumn = "mpgCat", x = mpg,
breaks = c(10, 17, 21, 35))
}
