% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCrossTables.R
\name{GenerateCrossTables}
\alias{GenerateCrossTables}
\title{Generate Cross Tables}
\usage{
GenerateCrossTables(Target, Covariate, Data, Groups, UseLogit = T)
}
\arguments{
\item{Target}{The name of the binary dependent variable.}

\item{Covariate}{The name of the covariates for which the cross tables
will be generated on.}

\item{Data}{A data.frame or data.table which contains the covariates and the
binary dependent variable.}

\item{Groups}{If the covariate is numeric, how many roughly equal-sized
groups should they be divided into.
This can be a vector of values. The highest value will first be attempted.
If roughly equal-sized groups cannot be created, the second highest value
will be attempted. If roughly equal-sized groups cannot be created from
the lowest value, the covariate will be treated as categorical with each
distinct value being a category.}

\item{UseLogit}{Argument for NumericalTable()/ CategoricalTable(). If the
value is TRUE, Log Odds will be generated. Otherwise, a set of score derived
from Log Odds, scaled from 0 to 100, will be generated.}
}
\description{
This function generates cross tables for the columns of covariates in a
data.table or data.frame in relation to a binary dependent variable. It
returns a list containing the cross tables
}
\examples{
TableList <- GenerateCrossTables(Target = "am",
Covariate = c("mpg", "cyl", "hp"), Groups = c(3, 4))
}
