% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateLogit.R
\name{GenerateLogit}
\alias{GenerateLogit}
\title{Generate Logit}
\usage{
GenerateLogit(DT, CrossTable, Covariate, Suffix = "Logit",
  SelectCol = NULL, NewName = NULL)
}
\arguments{
\item{CrossTable}{A cross table with the different levels of the covariate.}

\item{Covariate}{The name of the covariate.}

\item{Data}{A data.frame or data.table for which a new column will be
generated in.}

\item{UseLogit}{If the value is TRUE, the join will be based on a column
named "Logit". Otherwise, it will look for a column named "Score".}
}
\description{
This function generates a new column containing the conditional logit
corresponding to a covariate.
}
\examples{
Discretize a continuous covariate
mtcars[, mpgCat := cut(mpg, breaks = c(10, 17, 21, 35), include.lowest = T)]

Generate a cross table
mpgTable <- CategoricalTable("am", "mpgCat", mtcars)

Generate the conditional logit
GenerateLogit(mtcars, mpgTable, "mpgCat")
}
