% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{LogLoss}
\alias{LogLoss}
\title{Logarithmic Loss}
\usage{
LogLoss(Target, Prediction)
}
\arguments{
\item{Target}{A vector of the binary dependent variable.}

\item{Prediction}{A vector of the predictions.}
}
\description{
This function computes the logarithmic loss for binary predictions.
}
\examples{
LogLoss(Data[, Y], Model$fitted.values)
}
