% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericalTable.R
\name{NumericalTable}
\alias{NumericalTable}
\title{Cross Table of Numerical Covariates}
\usage{
NumericalTable(Target, Covariate, Data, NumberOfBins = 5,
  UseCustomIntervals = F, CustomIntervals = NULL, UseLogit = T)
}
\arguments{
\item{Target}{The name of binary target to be predicted.}

\item{Covariate}{The name of the covariate.}

\item{Data}{A data.table containing both the target and covariate.}

\item{NumberOfBins}{Number of bins the numerical value to be broken into.}

\item{UseCustomIntervals}{Allowing a custom set of values to be used for binning.}

\item{CustomIntervals}{The numerical values of the break points.}

\item{UseLogit}{If the value is TRUE, Log Odds will be generated. Otherwise,
a set of score derived from Log Odds, scaled from 0 to 100, will be generated.}
}
\description{
This function generates a data.table tabulating the different ranges of a continuous
covariate and the corresponding probabilities of the binary dependent
variable taking the value of 1.
}
\examples{
NumericalTable(Target = "am", Covariate = "mpg", Data = mtcars)
}
