% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{PlotROC}
\alias{PlotROC}
\title{Plot the ROC Curve}
\usage{
PlotROC(Target, Prediction)
}
\arguments{
\item{Target}{A vector with values of 0/1 to be predicted.}

\item{Prediction}{A vector of prediction or covariate used to predict the target.}
}
\description{
This function plots the ROC curve.
}
\examples{
PlotROC(mtcars$am, mtcars$mpg)
}
