% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Significance.R
\name{Significance}
\alias{Significance}
\title{Univariate Statistical Significance}
\usage{
Significance(Target, Prediction)
}
\arguments{
\item{Target}{A vector of the dependent variable.}

\item{Prediction}{A vector of the predictions.}
}
\description{
This function computes the Statistical Significance (p-value) equivalent to
that obtained from a linear regression.
}
\examples{
Significance(mtcars$mpg, mtcars$cyl)
}
