% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateSummary.R
\name{CovariateSummary}
\alias{CovariateSummary}
\title{Summary of Covariates}
\usage{
CovariateSummary(Target, Data, Skip = NULL, Type = "Binary")
}
\arguments{
\item{Target}{A string of the name of target to be predicted.}

\item{Data}{A data.frame or data.table containing both the target and
covariates.}

\item{Skip}{A vector of strings containing the names of columns to be
skipped, such as ID.}

\item{Type}{The type of dependent variables, either "Binary" or
"Continuous". Default is "Binary".}
}
\description{
This function allows you to explore the discriminatory power of covariates
against a binary target, and summarizes the number of NA and 0 in each of
the covariates.
}
\details{
If the Type is "Binary", it returns a data.table containing the names of
covariates, number of NA, number of 0, and Area Under the Curve (AUC).

If the Type is "Continuous", it returns a data.table containing the names of
covariates, number of NA, number of 0, Adjusted R-squared, and p-value.
}
\examples{
VariableSummary(Target = "am", Data = mtcars, Skip = c("vs", "carb"), Type = "Binary")
}
