% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateSummary.R
\name{CovariateSummary}
\alias{CovariateSummary}
\title{Summary of Covariates}
\usage{
CovariateSummary(Target, DT, Skip = NULL)
}
\arguments{
\item{Target}{A string of the name of target to be predicted.}

\item{DT}{A data.table containing both the target and
covariates.}

\item{Skip}{A vector of strings containing the names of columns to be
skipped.}
}
\description{
This function allows you to explore the discriminatory power of covariates
against a binary target, and summarizes the number of NA and 0 in each of
the covariates.
}
\details{
It returns a data.table containing the names of
covariates, number of NA, number of 0, and Area Under the Curve (AUC).
}
\examples{
CovariateSummary(Target = "am", DT = mtcars, Skip = c("vs", "carb"))
}
