% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{CovariateWeights}
\alias{CovariateWeights}
\title{Weights of Covariates}
\usage{
CovariateWeights(Model, Rounding = 0.01, Exact = FALSE, Intercept = TRUE)
}
\arguments{
\item{Model}{A "lm" or "glm" object.}

\item{Rounding}{Desired rounding, default is up to 2 decimal places.}

\item{Exact}{If Exact = TRUE, the function will return weights with the sum of exactly
100. Default is set to FALSE, to reduce time required for computation.}

\item{Intercept}{Does the regression object have an intercept term? Default is set to
TRUE.}
}
\description{
This function takes the regression object and calculates the relative
importance of different covariates.
The relative importance is only valid when the covariates are standardized.
}
\examples{
CovariateWeights(RegressionModel, Rounding = 1, Exact = TRUE)
}
