% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{GenerateLogit}
\alias{GenerateLogit}
\title{Generate Logit}
\usage{
GenerateLogit(Data, CrossTable, Covariate)
}
\arguments{
\item{Data}{A data.frame or data.table for which a new column will be
generated in.}

\item{CrossTable}{A cross table with the different levels of the covariate.}

\item{Covariate}{The name of the covariate.}
}
\description{
This function generates a new column containing the conditional logit
corresponding to a covariate.
}
\examples{
Discretize a continuous covariate
mtcars[, mpgCat := cut(mpg, breaks = c(10, 17, 21, 35), include.lowest = T)]

Generate the conditional logit
mtcars <- GenerateLogit(mtcars, mpgTable, "mpgCat")
}
