% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiteXploreR_functions.R
\name{NumericalTable}
\alias{NumericalTable}
\title{Cross Table of Numerical Covariates}
\usage{
NumericalTable(Target, Covariate, Data, NumberOfBins = 5, CustomBins = F,
  CustomIntervals = NULL)
}
\arguments{
\item{Target}{The name of binary target to be predicted.}

\item{Covariate}{The name of the covariate.}

\item{Data}{A data.table containing both the target and covariate.}
}
\description{
This function generates a data.table tabulating the different ranges of a continuous
covariate and the corresponding probabilities of the binary dependent
variable taking the value of 1.
}
\examples{
mpgTable <- NumericalTable(Target = "am", Covariate = "mpg", Data = mtcars)
}
