% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotROC.R
\name{PlotROC}
\alias{PlotROC}
\title{Plot ROC Curve}
\usage{
PlotROC(Target, Prediction, ...)
}
\arguments{
\item{Target}{A vector with values of 0/1 to be predicted.}

\item{Prediction}{A vector of prediction or covariate used to predict the target.}

\item{...}{Other parameters for plot().}
}
\description{
This function plots the ROC curve.
}
\examples{
PlotROC(mtcars$am, mtcars$mpg)
}
